refact: merge Transactions component into page/transactions
